chore: add command to build
