+++ packages/vscode-extension/package.json
@@ -552,6 +552,7 @@
     "typecheck": "tsc --noEmit",
     "format": "prettier --write --list-different src",
     "build:tests": "tsc --project tsconfig.test.json",
+    "build:plugins": "bash ./scripts/install-dev-plugins.sh",
     "test": "npm run build:extension && npm run build:webview && npm run build:tests && vscode-test",
     "preinstall": "cd ../vscode-js-debug && git submodule update --init ./ && npm install && npm exec tsc"
   },
