+++ packages/vscode-extension/plugins_dist/redux/foldgutter-1f00b44a780329dcb1a9cb6f53c74899.css
@@ -1,0 +1,1 @@
+.CodeMirror-foldmarker{color:#00f;text-shadow:1px 1px 2px #b9f,-1px -1px 2px #b9f,1px -1px 2px #b9f,-1px 1px 2px #b9f;cursor:pointer;font-family:arial;line-height:.3}.CodeMirror-foldgutter{width:.7em}.CodeMirror-foldgutter-open,.CodeMirror-foldgutter-folded{cursor:pointer}.CodeMirror-foldgutter-open:after{content:"▾"}.CodeMirror-foldgutter-folded:after{content:"▸"}
