+++ packages/vscode-extension/src/plugins/redux-devtools-plugin/ReduxDevToolsPluginWebviewProvider.ts
@@ -1,3 +1,4 @@
+import fs from "fs";
 import {
   CancellationToken,
   ExtensionContext,
@@ -9,13 +10,37 @@ import {
 } from "vscode";
 import { extensionContext } from "../../utilities/extensionContext";
 import { getUri } from "../../utilities/getUri";
+import { getNonce } from "../../utilities/getNonce";
+
+const PATH = "plugins_dist/redux/";
+
+const prepareWebviewCSS = (files: string[]) => {
+  return files.map((file) => {
+    return /*html*/ `
+      <link rel="preload" href="${file}" as="style">
+      <link rel="stylesheet" href="${file}">
+    `;
+  }).join("\n");
+};
+
+const prepareWebviewJS = (files: string[], nonce: string) => {
+  return files.map((file) => {
+    return /*html*/ `<script src="${file}" nonce="${nonce}" defer></script>`;
+  }).join("\n");
+};
 
 function generateWebviewContent(
   context: ExtensionContext,
   webview: Webview,
   extensionUri: Uri
 ): string {
-  const baseUri = getUri(webview, extensionUri, ["dist/plugins/redux"]);
+  const nonce = getNonce();
+  const baseUri = getUri(webview, extensionUri, [PATH]);
+  const files = fs.readdirSync(baseUri.path, {recursive: true});
+  const cssFiles = files.filter((file) => typeof file === 'string').filter((file) => file.endsWith(".css"));
+  const jsFiles = files.filter((file) => typeof file === 'string').filter((file) => file.endsWith(".js"));
+  const cssImports = prepareWebviewCSS(cssFiles);
+  const jsImports = prepareWebviewJS(jsFiles, nonce);
 
   return /*html*/ `
     <!DOCTYPE html>
@@ -44,13 +69,16 @@ function generateWebviewContent(
               flex: 1;
             }
           </style>
-        
-          <link rel="preload" href="./redux/css/index-8f30c04047bae6699f82f40acebe1aed.css" as="style">
-          <link rel="stylesheet" href="./redux/css/index-8f30c04047bae6699f82f40acebe1aed.css">
-          <link rel="preload" href="./redux/css/codemirror-46292010b15a8ba62a25f29c89e1e049.css" as="style">
-          <link rel="stylesheet" href="./redux/css/codemirror-46292010b15a8ba62a25f29c89e1e049.css">
-          <link rel="preload" href="./redux/css/foldgutter-1f00b44a780329dcb1a9cb6f53c74899.css" as="style">
-          <link rel="stylesheet" href="./redux/css/foldgutter-1f00b44a780329dcb1a9cb6f53c74899.css"></head>
+          <meta http-equiv="Content-Security-Policy"
+            content="default-src 'none';
+            img-src vscode-resource: http: https: data:;
+            media-src vscode-resource: http: https:;
+            style-src ${webview.cspSource} 'unsafe-inline';
+            script-src 'nonce-${nonce}';
+            font-src vscode-resource: https:;" 
+          />
+          ${cssImports}
+        </head>
 
         <body>
           <!-- Use static rendering with Expo Router to support running without JavaScript. -->
@@ -59,7 +87,7 @@ function generateWebviewContent(
           </noscript>
           <!-- The root element for your Expo app. -->
           <div id="root"></div>
-        <script src="./redux/js/AppEntry-9f5185ef27313606ee5279432a7a2e89.js" defer></script>
+          ${jsImports}
       </body>
       </html>
   `;
@@ -77,7 +105,7 @@ export class ReduxDevToolsPluginWebviewProvider implements WebviewViewProvider {
   ): void {
     webviewView.webview.options = {
       enableScripts: true,
-      localResourceRoots: [Uri.joinPath(this.context.extensionUri, "dist")],
+      localResourceRoots: [Uri.joinPath(this.context.extensionUri, PATH)],
     };
 
     webviewView.webview.html = generateWebviewContent(
